Make rendering the type help in the pager work.
