+++ CMakeLists.txt
@@ -136,7 +136,7 @@ include(CheckCXXCompilerFlag)
 
 string(TOUPPER "${CMAKE_BUILD_TYPE}" U_CMAKE_BUILD_TYPE)
 
-set(CMAKE_INSTALL_RPATH "${CMAKE_INSTALL_PREFIX}/lib; ${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR}")
+#set(CMAKE_INSTALL_RPATH "${CMAKE_INSTALL_PREFIX}/lib; ${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR}")
 
 macro(xeus_cpp_set_common_options target_name)
     if (MSVC)
@@ -161,7 +161,7 @@ macro(xeus_cpp_set_common_options target_name)
 
     if (APPLE)
         set_target_properties(${target_name} PROPERTIES
-            MACOSX_RPATH ON
+            MACOSX_RPATH OFF
         )
     else ()
         set_target_properties(${target_name} PROPERTIES
@@ -171,7 +171,7 @@ macro(xeus_cpp_set_common_options target_name)
     endif ()
 
     set_target_properties(${target_name} PROPERTIES
-        INSTALL_RPATH_USE_LINK_PATH TRUE
+        INSTALL_RPATH_USE_LINK_PATH FALSE
     )
 endmacro()
