+++ etc/xeus-cpp/tags.d/stl.json
@@ -1,0 +1,4 @@
+{
+    "url": "https://en.cppreference.com/w/",
+    "tagfile": "cppreference-doxygen-web.tag"
+}
