+++ etc/xeus-cpp/tags.d/xtensor.json
@@ -1,0 +1,4 @@
+{
+    "url": "https://xtensor.readthedocs.io/en/latest/",
+    "tagfile": "xtensor.tag"
+}
