+++ src/xinspect.hpp
@@ -1,4 +1,4 @@
-*/***********************************************************************************
+/***********************************************************************************
  * Copyright (c) 2023, xeus-cpp contributors                                        *
  * Copyright (c) 2023, Johan Mabille, Loic Gouarin, Sylvain Corlay, Wolf Vollprecht *
  *                                                                                  *
@@ -6,7 +6,6 @@
  *                                                                                  *
  * The full license is in the file LICENSE, distributed with this software.         *
  ************************************************************************************/
-
 #ifndef XEUS_CPP_INSPECT_HPP
 #define XEUS_CPP_INSPECT_HPP
 
@@ -14,10 +13,6 @@
 #include <string>
 
 #include <dirent.h>
-
-#include "cling/Interpreter/Interpreter.h"
-#include "cling/Interpreter/Value.h"
-#include "cling/Utils/Output.h"
 
 #include <pugixml.hpp>
 
@@ -84,62 +79,23 @@ namespace xcpp
         }
     };
 
-    /*    std::string find_type(const std::string& expression, cling::Interpreter& interpreter)
-        {
-            cling::Value result;
-            std::string typeString;
-
-            // add typeinfo in include files in order to use typeid
-            std::string code = "#include <typeinfo>";
-            auto compilation_result = interpreter.process(code.c_str(), &result);
-
-            // try to find the typename of the class
-            code = "typeid(" + expression + ").name();";
-
-            // Temporarily dismissing all std::cerr and std::cout resulting from `interpreter.process`
-            compilation_result = interpreter.process(code.c_str(), &result);
-            {
-                auto cout_strbuf = std::cout.rdbuf();
-                auto cerr_strbuf = std::cerr.rdbuf();
-                auto null = xnull();
-                std::cout.rdbuf(&null);
-                std::cerr.rdbuf(&null);
-
-                compilation_result = interpreter.process(code.c_str(), &result);
-
-                std::cout.rdbuf(cout_strbuf);
-                std::cerr.rdbuf(cerr_strbuf);
-            }
-
-            if (compilation_result == cling::Interpreter::kSuccess)
-            {
-                // we found the typeid
-                std::string valueString;
-                {
-                    llvm::raw_string_ostream os(valueString);
-                    result.print(os);
-                }
-
-                // search the typename in the output between ""
-                std::regex re_typename("\\\"(.*)\\\"");
-                std::smatch typename_;
-                std::regex_search(valueString, typename_, re_typename);
-                // set in valueString the typename given by typeid
-                valueString = typename_.str(1);
-                // we need demangling in order to have its string representation
-                valueString = demangle(valueString);
-
-                re_typename = "(\\w*(?:\\:{2}?\\w*)*)";
-                std::regex_search(valueString, typename_, re_typename);
-                if (!typename_.str(1).empty())
-                {
-                    typeString = typename_[1];
-                }
-            }
-
-            return typeString;
+    std::string find_type(const std::string& expression, clang::Interpreter& interpreter)
+    {
+        auto PTU = interpreter.Parse(expression + ";");
+        if (llvm::Error Err = PTU.takeError()) {
+            llvm::logAllUnhandledErrors(std::move(Err), llvm::errs(), "error: ");
+            return "";
         }
-        */
+
+	    clang::Decl *D = *PTU->TUPart->decls_begin();
+	    if (!llvm::isa<clang::TopLevelStmtDecl>(D))
+	        return "";
+
+	    clang::Expr *E = llvm::cast<clang::Expr>(llvm::cast<clang::TopLevelStmtDecl>(D)->getStmt());
+
+	    clang::QualType QT = E->getType();
+        return  QT.getAsString();
+    }
 
     static nl::json read_tagconfs(const char* path)
     {
@@ -171,178 +127,152 @@ namespace xcpp
         return result;
     }
 
-    void inspect(const std::string& code, nl::json& kernel_res, cling::Interpreter& interpreter)
+    std::pair<bool, std::smatch> is_inspect_request(const std::string code, std::regex re) 
     {
-        /*
-                std::string tagconf_dir = XCPP_TAGCONFS_DIR;
-                std::string tagfiles_dir = XCPP_TAGFILES_DIR;
+        std::smatch inspect;
+        if (std::regex_search(code, inspect, re)){
+            return std::make_pair(true, inspect);
+        }
+        return std::make_pair(false, inspect);
+       
+    }
 
-                nl::json tagconfs = read_tagconfs(tagconf_dir.c_str());
+    void inspect(const std::string& code, nl::json& kernel_res, clang::Interpreter& interpreter)
+    {
+        std::string tagconf_dir = XCPP_TAGCONFS_DIR;
+        std::string tagfiles_dir = XCPP_TAGFILES_DIR;
 
-                std::vector<std::string> check{"class", "struct", "function"};
+        nl::json tagconfs = read_tagconfs(tagconf_dir.c_str());
 
-                std::string url, tagfile;
+        std::vector<std::string> check{"class", "struct", "function"};
 
-                std::regex re_expression(R"((((?:\w*(?:\:{2}|\<.*\>|\(.*\)|\[.*\])?)\.?)*))");
-                std::smatch inspect;
-                std::regex_search(code, inspect, re_expression);
+        std::string url, tagfile;
 
-                std::string inspect_result;
+        std::regex re_expression(R"((((?:\w*(?:\:{2}|\<.*\>|\(.*\)|\[.*\])?)\.?)*))");
 
-                std::smatch method;
-                std::string to_inspect = inspect[1];
+        std::smatch inspect = is_inspect_request(code, re_expression).second;
 
-                // Method or variable of class found (xxxx.yyyy)
-                if (std::regex_search(to_inspect, method, std::regex(R"((.*)\.(\w*)$)")))
+        std::string inspect_result;
+
+        std::smatch method;
+        std::string to_inspect = inspect[1];
+
+        // Method or variable of class found (xxxx.yyyy)
+        if (std::regex_search(to_inspect, method, std::regex(R"((.*)\.(\w*)$)")))
+        {
+            std::string typename_ = find_type(method[1], interpreter);
+
+            if (!typename_.empty())
+            {
+                for (nl::json::const_iterator it = tagconfs.cbegin(); it != tagconfs.cend(); ++it)
                 {
-                    std::string typename_ = find_type(method[1], interpreter);
-
-                    if (!typename_.empty())
+                    url = it->at("url");
+                    tagfile = it->at("tagfile");
+                    std::string filename = tagfiles_dir + "/" + tagfile;
+                    pugi::xml_document doc;
+                    pugi::xml_parse_result result = doc.load_file(filename.c_str());
+                    class_member_predicate predicate{typename_, "function", method[2]};
+                    auto node = doc.find_node(predicate);
+                    if (!node.empty())
                     {
-                        for (nl::json::const_iterator it = tagconfs.cbegin(); it != tagconfs.cend(); ++it)
-                        {
-                            url = it->at("url");
-                            tagfile = it->at("tagfile");
-                            std::string filename = tagfiles_dir + "/" + tagfile;
-                            pugi::xml_document doc;
-                            pugi::xml_parse_result result = doc.load_file(filename.c_str());
-                            class_member_predicate predicate{typename_, "function", method[2]};
-                            auto node = doc.find_node(predicate);
-                            if (!node.empty())
-                            {
-                                inspect_result = url + predicate.get_filename(node);
-                            }
-                        }
+                        inspect_result = url + predicate.get_filename(node);
                     }
                 }
-                else
-                {
-                    std::string find_string;
+            }
+        }
+        else
+        {
+            std::string find_string;
 
-                    // check if we try to find the documentation of a namespace
-                    // if yes, don't try to find the type using typeid
-                    std::regex is_namespace(R"(\w+(\:{2}\w+)+)");
-                    std::smatch namespace_match;
-                    if (std::regex_match(to_inspect, namespace_match, is_namespace))
+            // check if we try to find the documentation of a namespace
+            // if yes, don't try to find the type using typeid
+            std::regex is_namespace(R"(\w+(\:{2}\w+)+)");
+            std::smatch namespace_match;
+            if (std::regex_match(to_inspect, namespace_match, is_namespace))
+            {
+                find_string = to_inspect;
+            }
+            else
+            {
+                std::string typename_ = find_type(to_inspect, interpreter);
+                find_string = (typename_.empty()) ? to_inspect : typename_;
+            }
+
+            for (nl::json::const_iterator it = tagconfs.cbegin(); it != tagconfs.cend(); ++it)
+            {
+                url = it->at("url");
+                tagfile = it->at("tagfile");
+                std::string filename = tagfiles_dir + "/" + tagfile;
+                pugi::xml_document doc;
+                pugi::xml_parse_result result = doc.load_file(filename.c_str());
+                for (auto c : check)
+                {
+                    node_predicate predicate{c, find_string};
+                    std::string node;
+
+                    if (c == "class" || c == "struct")
                     {
-                        find_string = to_inspect;
+                        node = doc.find_node(predicate).child("filename").child_value();
                     }
                     else
                     {
-                        std::string typename_ = find_type(to_inspect, interpreter);
-                        find_string = (typename_.empty()) ? to_inspect : typename_;
+                        node = doc.find_node(predicate).child("anchorfile").child_value();
                     }
 
-                    for (nl::json::const_iterator it = tagconfs.cbegin(); it != tagconfs.cend(); ++it)
+                    if (!node.empty())
                     {
-                        url = it->at("url");
-                        tagfile = it->at("tagfile");
-                        std::string filename = tagfiles_dir + "/" + tagfile;
-                        pugi::xml_document doc;
-                        pugi::xml_parse_result result = doc.load_file(filename.c_str());
-                        for (auto c : check)
-                        {
-                            node_predicate predicate{c, find_string};
-                            std::string node;
-
-                            if (c == "class" || c == "struct")
-                            {
-                                node = doc.find_node(predicate).child("filename").child_value();
-                            }
-                            else
-                            {
-                                node = doc.find_node(predicate).child("anchorfile").child_value();
-                            }
-
-                            if (!node.empty())
-                            {
-                                inspect_result = url + node;
-                            }
-                        }
+                        inspect_result = url + node;
                     }
                 }
+            }
+        }
 
-                if (inspect_result.empty())
-                {
-                    std::cerr << "No documentation found for " << code << "\n";
-                    std::cout << std::flush;
-                    kernel_res["found"] = false;
-                    kernel_res["status"] = "error";
-                    kernel_res["ename"] = "No documentation found";
-                    kernel_res["evalue"] = "";
-                    kernel_res["traceback"] = nl::json::array();
-                }
-                else
-                {
-                    // Format html content.
-                    std::string html_content = R"(<style>
-                    #pager-container {
-                        padding: 0;
-                        margin: 0;
-                        width: 100%;
-                        height: 100%;
-                    }
-                    .xcpp-iframe-pager {
-                        padding: 0;
-                        margin: 0;
-                        width: 100%;
-                        height: 100%;
-                        border: none;
-                    }
-                    </style>
-                    <iframe class="xcpp-iframe-pager" src=")" +
-                        inspect_result + R"(?action=purge"></iframe>)";
+        if (inspect_result.empty())
+        {
+            std::cerr << "No documentation found for " << code << "\n";
+            std::cout << std::flush;
+            kernel_res["found"] = false;
+            kernel_res["status"] = "error";
+            kernel_res["ename"] = "No documentation found";
+            kernel_res["evalue"] = "";
+            kernel_res["traceback"] = nl::json::array();
+        }
+        else
+        {
+            // Format html content.
+            std::string html_content = R"(<style>
+            #pager-container {
+                padding: 0;
+                margin: 0;
+                width: 100%;
+                height: 100%;
+            }
+            .xcpp-iframe-pager {
+                padding: 0;
+                margin: 0;
+                width: 100%;
+                height: 100%;
+                border: none;
+            }
+            </style>
+            <iframe class="xcpp-iframe-pager" src=")"
+                                       + inspect_result + R"(?action=purge"></iframe>)";
 
-                    // Note: Adding "?action=purge" suffix to force cppreference's
-                    // Mediawiki to purge the HTTP cache.
+            // Note: Adding "?action=purge" suffix to force cppreference's
+            // Mediawiki to purge the HTTP cache.
 
-                    kernel_res["payload"] = nl::json::array();
-                    kernel_res["payload"][0] = nl::json::object({
-                        {"data", {
-                            {"text/plain", inspect_result},
-                            {"text/html", html_content}}
-                        },
-                        {"source", "page"},
-                        {"start", 0}
-                    });
-                    kernel_res["user_expressions"] = nl::json::object();
+            kernel_res["payload"] = nl::json::array();
+            kernel_res["payload"][0] = nl::json::object(
+                {{"data", {{"text/plain", inspect_result}, {"text/html", html_content}}},
+                 {"source", "page"},
+                 {"start", 0}}
+            );
+            kernel_res["user_expressions"] = nl::json::object();
 
-                    std::cout << std::flush;
-                    kernel_res["found"] = true;
-                    kernel_res["status"] = "ok";
-                }
-            */
+            std::cout << std::flush;
+            kernel_res["found"] = true;
+            kernel_res["status"] = "ok";
+        }
     }
-
-    class xintrospection : public xpreamble
-    {
-    public:
-
-        using xpreamble::pattern;
-        const std::string spattern = R"(^\?)";
-
-        xintrospection(cling::Interpreter& p)
-            : m_interpreter{p}
-        {
-            pattern = spattern;
-        }
-
-        void apply(const std::string& code, nl::json& kernel_res) override
-        {
-            std::regex re(spattern + R"((.*))");
-            std::smatch to_inspect;
-            std::regex_search(code, to_inspect, re);
-            inspect(to_inspect[1], kernel_res, m_interpreter);
-        }
-
-        virtual xpreamble* clone() const override
-        {
-            return new xintrospection(*this);
-        }
-
-    private:
-
-        cling::Interpreter& m_interpreter;
-    };
 }
 #endif
